Move changes to nightly
Remove incorrect UTC run comment

rename back to nightly

Totally not doing two things at once

Pretend utc still works

normal utc then?

clarify a bit
