--- .github/workflows/deploy-single.yaml
@@ -1,38 +1,0 @@
-name: Deploy manually
-on: workflow_dispatch
-env:
-  LOGFILE_NAME: full_output.log
-  ARCHIVE_SUBPROCESS_LOG: 7z_output.log  
-jobs:
-  manually:
-    name: Deploy manually
-    runs-on: ubuntu-latest
-    permissions:
-      contents: write
-
-    steps:
-      # https://github.com/actions/checkout
-      - name: Checkout repository
-        uses: actions/checkout@v6
-        with:
-          fetch-depth: 0
-      - name: Install dependencies
-        run: pip install PyGithub GitPython
-      - name: Deploy manually
-        env:
-          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
-          ADJUSTED_DATE_SINCE: 2025-05-01
-          DRY_RUN: False
-          SINCE_RELEASE: False
-          COMPRESSION_TIMEOUT: 120 # seconds until 7z subprocess is killed
-        run: python ".github/workflows/deploy-single.py"
-      # https://github.com/actions/upload-artifact
-      - name: Store log as artifact
-        uses: actions/upload-artifact@v7
-        if: always()
-        with: 
-          name: logs-from-job-run-${{ github.run_number }}
-          retention-days: 7
-          path: |
-            ${{ env.LOGFILE_NAME }}
-            ${{ env.ARCHIVE_SUBPROCESS_LOG }}
